Refactoring test folder
